+++ Sketch/Eyes.h
@@ -1,70 +1,65 @@
 //Led matrix - eyes
-const short idleEyes[6][2]={
+const uint8_t idleEyes[6][2]={
   {7,5},{8,5},{9,5},
   {7,10},{8,10},{9,10}
 };
 
-const short happyCenterEyes[6][2] = {
+const uint8_t happyCenterEyes[6][2] = {
   {7,5},{8,5},{9,5},
   {7,10},{8,10},{9,10}
 };
 
-const short happyLeftEyes[6][2] = {
+const uint8_t happyLeftEyes[6][2] = {
   {7,4},{8,4},{9,4},
   {7,10},{8,10},{9,10}
 };
 
-const short happyRightEyes[6][2] = {
+const uint8_t happyRightEyes[6][2] = {
   {7,5},{8,5},{9,5},
   {7,11},{8,11},{9,11}
 };
 
-
-const short sadCenterEyes[6][2] = {
+const uint8_t sadCenterEyes[6][2] = {
   {7,5},{8,5},{9,5},
   {7,10},{8,10},{9,10}
 };
 
-const short sadLeftEyes[6][2] = {
+const uint8_t sadLeftEyes[6][2] = {
   {7,4},{8,4},{9,4},
   {7,10},{8,10},{9,10}
 };
 
-const short sadRightEyes[6][2] = {
+const uint8_t sadRightEyes[6][2] = {
   {7,5},{8,5},{9,5},
   {7,11},{8,11},{9,11}
 };
 
-const short angryCenterEyes[6][2] = {
+const uint8_t angryCenterEyes[6][2] = {
   {7,5},{8,5},{9,5},
   {7,10},{8,10},{9,10}
 };
 
-const short angryLeftEyes[6][2] = {
+const uint8_t angryLeftEyes[6][2] = {
   {7,4},{8,4},{9,4},
   {7,10},{8,10},{9,10}
 };
 
-const short angryRightEyes[6][2] = {
+const uint8_t angryRightEyes[6][2] = {
   {7,5},{8,5},{9,5},
   {7,11},{8,11},{9,11}
 };
 
-
-const short scaredCenterEyes[4][2] = {
+const uint8_t scaredCenterEyes[4][2] = {
   {5,5},{6,5},
   {5,10},{6,10}
 };
 
-const short scaredLeftEyes[4][2] = {
+const uint8_t scaredLeftEyes[4][2] = {
   {5,4},{6,4},
   {5,10},{6,10}
 };
 
-const short scaredRightEyes[4][2] = {
+const uint8_t scaredRightEyes[4][2] = {
   {5,5},{6,5},
   {5,11},{6,11}
 };
-
-
-
