+++ Sketch/LedStrips.ino
@@ -19,51 +19,51 @@ void drawFixedStrips() {
 
   switch (moodState) {
     case happy:
-      for (int i = 0; i < bottomStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < bottomStrip.numPixels(); i = i + 3) {
         bottomStrip.setPixelColor(i, bottomStrip.Color(GREEN));
 
       }
-      for (int i = 0; i < topStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < topStrip.numPixels(); i = i + 3) {
         topStrip.setPixelColor(i, topStrip.Color(GREEN));
       }
       break;
 
     case sad:
-      for (int i = 0; i < bottomStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < bottomStrip.numPixels(); i = i + 3) {
         bottomStrip.setPixelColor(i, bottomStrip.Color(PURPLE));
 
       }
-      for (int i = 0; i < topStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < topStrip.numPixels(); i = i + 3) {
         topStrip.setPixelColor(i, topStrip.Color(PURPLE));
       }
       break;
 
     case angry:
-      for (int i = 0; i < bottomStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < bottomStrip.numPixels(); i = i + 3) {
         bottomStrip.setPixelColor(i, bottomStrip.Color(RED));
 
       }
-      for (int i = 0; i < topStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < topStrip.numPixels(); i = i + 3) {
         topStrip.setPixelColor(i, topStrip.Color(RED));
       }
       break;
 
     case scared:
-      for (int i = 0; i < bottomStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < bottomStrip.numPixels(); i = i + 3) {
         bottomStrip.setPixelColor(i, bottomStrip.Color(BLUE));
 
       }
-      for (int i = 0; i < topStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < topStrip.numPixels(); i = i + 3) {
         topStrip.setPixelColor(i, topStrip.Color(BLUE));
       }
       break;
 
     default:
-      for (int i = 0; i < bottomStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < bottomStrip.numPixels(); i = i + 3) {
         bottomStrip.setPixelColor(i, bottomStrip.Color(YELLOW));
 
       }
-      for (int i = 0; i < topStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < topStrip.numPixels(); i = i + 3) {
         topStrip.setPixelColor(i, topStrip.Color(YELLOW));
       }
   }
@@ -75,7 +75,7 @@ void drawFixedStrips() {
 //Turn every third RGB led on in a way that it seems Teo is breathing
 //Obs: you must not use setBrightness method as animation. Use colorPercentage of setPixelColor method
 void drawBreathingStrips() {
-  int red, green, blue = 0;
+  uint8_t red, green, blue = 0;
 
   switch (moodState) {
     case happy:
@@ -116,10 +116,10 @@ void drawBreathingStrips() {
 
   if (isTeoBreathingOut == true) {
     for (int colorPercentage = 0; colorPercentage <= 255; colorPercentage = colorPercentage + 2) {
-      for (int i = 0; i < bottomStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < bottomStrip.numPixels(); i = i + 3) {
         bottomStrip.setPixelColor(i, bottomStrip.Color(colorPercentage * red, colorPercentage * green, colorPercentage * blue));
       }
-      for (int i = 0; i < topStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < topStrip.numPixels(); i = i + 3) {
         topStrip.setPixelColor(i, topStrip.Color(colorPercentage * red, colorPercentage * green, colorPercentage * blue));
       }
       bottomStrip.show();
@@ -130,10 +130,10 @@ void drawBreathingStrips() {
 
   } else {
     for (int colorPercentage = 255; colorPercentage >= 0; colorPercentage = colorPercentage - 2) {
-      for (int i = 0; i < bottomStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < bottomStrip.numPixels(); i = i + 3) {
         bottomStrip.setPixelColor(i, bottomStrip.Color(colorPercentage * red, colorPercentage * green, colorPercentage * blue)); //turn every third pixel on
       }
-      for (int i = 0; i < topStrip.numPixels(); i = i + 3) {
+      for (uint8_t i = 0; i < topStrip.numPixels(); i = i + 3) {
         topStrip.setPixelColor(i, topStrip.Color(colorPercentage * red, colorPercentage * green, colorPercentage * blue)); //turn every third pixel on
       }
       bottomStrip.show();
@@ -144,4 +144,3 @@ void drawBreathingStrips() {
 
   }
 }
-
